+++ numba_dpex/core/parfors/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: 2020 - 2023 Intel Corporation
+#
+# SPDX-License-Identifier: Apache-2.0
