+++ numba_dpex/core/parfors/reduction_helper.py
@@ -296,7 +296,7 @@ class ReductionKernelVariables:
 
         param_types = [_to_scalar_from_0d(typemap[v]) for v in parfor_params]
 
-        # Calculate types of args passed to gufunc.
+        # Calculate types of args passed to the kernel function.
         func_arg_types = [typemap[v] for v in (parfor_inputs + parfor_outputs)]
 
         # Replace illegal parameter names in the loop body with legal ones.
